requirements.txt
# Core
python-dotenv
asyncio

# Web Framework
fastapi
uvicorn[standard]
websockets
aiohttp
pydantic

# Database
sqlalchemy
databases[postgresql]
psycopg2-binary
alembic

# Timezone
pytz

# AI & LangChain with Ollama
langchain==0.1.0
langchain-core
langchain-community
ollama  # Ollama Python 클라이언트
chromadb  # 벡터 DB (FAISS 대체)
sentence-transformers  # 로컬 임베딩
pypdf  # PDF 처리
pillow  # 이미지 처리
pytesseract  # OCR (이미지 텍스트 추출)

# Social Media APIs
requests
facebook-sdk
instagrapi
python-twitter

# Email
aiosmtplib
email-validator

# Utils
python-multipart
python-jose[cryptography]
passlib[bcrypt]
redis
celery

python-dotenv
httpx
websockets
jinja2
fastapi
SQLAlchemy
pydantic